User-Call-Automation/Agent-voice-call/requirements.txt
google-generativeai>=0.3.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
python-dotenv>=1.0.0
mangum>=0.17.0
